backend/llm_utils.py
from langchain_core.prompts import PromptTemplate
import google.generativeai as genai
import os
from dotenv import load_dotenv

load_dotenv()
GOOGLE_API_KEY = os.getenv("GOOGLE_API_KEY")
genai.configure(api_key=GOOGLE_API_KEY)

gemini_model = genai.GenerativeModel("gemini-1.5-flash")

prompt_template = PromptTemplate(
    template="""
      You are a helpful assistant.
      Answer ONLY from the provided transcript context.
      If the context is insufficient, just say you don't know.

      {context}
      Question: {question}
    """,
    input_variables=['context', 'question']
)

def generate_answer(context: str, question: str) -> str:
    prompt = prompt_template.format(context=context, question=question)
    response = gemini_model.generate_content(prompt)
    return response.text
