backend/requirements.txt
fastapi
uvicorn
youtube-transcript-api==1.2.2
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
